feat: Add option to disable the subtitles
By defaut the subtitles will be disabled instead of selecting the first
subtitle text stream extracted from the file.

In the subtitles drop down the first item will be the `Disabled` option
while the extracted subtitles will be added after it.

This first commit introduces some duplicated code when it comes to
reading and setting the pipeline flags. I'll try to remove it, but for
now this is a good starting point.